+++ Escape/Assets/Scripts/HealthBar.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class HealthBar : MonoBehaviour{
     public Slider slider;
 
-    public void SetHealth(int health){
+    public void SetHealth(float health){
         slider.value = health;
     }
     public void SetMaxHealth(int health){
